Fork log/syslog in order to support custom tags.
The current log/syslog package does not allow modification of the syslog
tag after a network connection is dialed. Normally this is okay, but since
logspout monitors several different processes and forwards them, we also
need to tag each process.

Additionally, tag and priority are public fields now, so if there's a need to
change the default tag and priority, it can be done.

Also, dialing every time we need to send something to syslog sucks.

Signed-off-by: Robert Xu <robxu9@gmail.com>