make it easier to build the docker image
i don't want to have to install Go, setup $GOPATH, install Go deps ... all on
the host, just to produce a binary to include in the docker image. building
should happen within the docker image so that host can be left intact.

to build and produce the image, merely run:

  make