router/pump: remove logstream send timeout
Previously a containers logstream would be closed after a send on
its logstream channel was blocked for 1 second, causing it to not
forward logs until logspout received a start event for the container.

A send to a logstream could happen due to a closed syslog/tcp connection
or any delay in an adapters Stream from processing incoming messages.

Although this commit fixes the current issue, it does open up the
possibility for a send on a logstream to block for an indefinite amount
of time.