Added retry to raw+tcp adapter
Most of this change is duplicated from the syslog adapter, since it
contains a working retry. Since the raw adapter also doesn't have
any tests associated with it, this is not implemented into this
version.

Since the documentation already contains the necessary retry
information for TCP adapters, I didn't see any need to alter this.