Add optional TCP framing to syslog adapter
Enable users to specify "octet-counted" frames as described in RFC6587
(Syslog over TCP) 3.4.1 and RFC5424 (Syslog over TLS). This prefixes
each message with the length of the message to allow consumers to easily
determine where the message ends (rather than traditional LF framing).
This also enables multiline Syslog messages without escaping.

This keeps the default as LF framing for backwards compatibility though
octet-counted framing is preferred when it is known the downstream
consumer can handle it.